(T): Split tests upon corresponding target-types to have an ability      set module-type properly
